requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.5
django-admin-sortable2==2.1.4
django-grappelli==3.0.4
django-tinymce==3.5.0
sqlparse==0.4.3
tzdata==2022.7
